AllControllers and Repositories Done

+++ Controllers/ReturnRequestsController.cs
@@ -18,11 +18,17 @@ namespace Hexa_Hub.Controllers
     {
         private readonly IReturnReqRepo _returnRequestRepo;
         private readonly DataContext _context;
+        private readonly IAsset _asset;
+        private readonly IAssetAllocation _assetAlloc;
+        private readonly IAssetRequest _assetRequest;
 
-        public ReturnRequestsController(DataContext context, IReturnReqRepo returnRequestRepo)
+        public ReturnRequestsController(DataContext context, IReturnReqRepo returnRequestRepo,IAsset asset,IAssetAllocation assetAllocation,IAssetRequest assetRequest)
         {
             _context = context;
             _returnRequestRepo = returnRequestRepo;
+            _asset = asset;
+            _assetAlloc = assetAllocation;
+            _assetRequest = assetRequest;
         }
 
         // GET: api/ReturnRequests
@@ -33,18 +39,18 @@ namespace Hexa_Hub.Controllers
             //return await _context.ReturnRequests.ToListAsync();
             var userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
             var userRole = User.FindFirstValue(ClaimTypes.Role);
-            if(userRole == "Admin")
+            if (userRole == "Admin")
             {
                 return await _returnRequestRepo.GetAllReturnRequest();
             }
             else
             {
-                var req = await _returnRequestRepo.GetReturnRequestById(userId);
-                if (req == null)
+                var req = await _returnRequestRepo.GetReturnRequestsByUserId(userId);
+                if (req == null||req.Count==0)
                 {
-                    return NotFound();
+                    return NotFound($"No details found");
                 }
-                return Ok(new List<ReturnRequest> { req });
+                return Ok(req);
             }
         }
 
@@ -53,65 +59,91 @@ namespace Hexa_Hub.Controllers
         [Authorize]
         public async Task<ActionResult<ReturnRequest>> GetReturnRequest(int id)
         {
-            //var returnRequest = await _context.ReturnRequests.FindAsync(id);
-            //var returnRequest = await _returnRequestRepo.GetReturnRequestById(id);
-
-            //if (returnRequest == null)
-            //{
-            //    return NotFound();
-            //}
-
-            //return returnRequest;
             var userRole = User.FindFirstValue(ClaimTypes.Role);
             if (userRole == "Admin")
             {
                 var req = await _returnRequestRepo.GetReturnRequestById(id);
-                if (req == null)
+                if(req == null)
                 {
-                    return NotFound();
+                    return NotFound($"Details For the User id {id} is not found");
                 }
-                return req;
+                return Ok(req);
             }
-            else
-            {
-                var userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
-                var req = await _returnRequestRepo.GetReturnRequestById(userId);
-                if (req == null)
-                {
-                    return NotFound();
-                }
-                return Ok(new List<ReturnRequest> { req });
-            }
+            return Forbid();
         }
+
 
         // PUT: api/ReturnRequests/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        [Authorize(Roles = "User")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> PutReturnRequest(int id, ReturnRequest returnRequest)
         {
             if (id != returnRequest.ReturnId)
             {
                 return BadRequest();
             }
-
-            var userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
-            var existingId = await _returnRequestRepo.GetReturnRequestById(id);
-            if (existingId == null)
+            var exisitingRequest = await _returnRequestRepo.GetReturnRequestById(id);
+            if (exisitingRequest == null)
             {
                 return NotFound();
             }
-            if (existingId.UserId != userId && !User.IsInRole("Admin"))
-            {
-                return Forbid();
-            }
-            //_context.Entry(returnRequest).State = EntityState.Modified;
-            _returnRequestRepo.UpdateReturnRequest(returnRequest);
+            _context.Entry(exisitingRequest).CurrentValues.SetValues(returnRequest);
 
+            if (returnRequest.ReturnStatus == Models.MultiValues.ReturnReqStatus.Approved)
+            {
+                exisitingRequest.ReturnDate = DateTime.Now;
+            }
+            if (returnRequest.ReturnStatus == Models.MultiValues.ReturnReqStatus.Returned)
+            {
+                exisitingRequest.ReturnDate = DateTime.Now;
+                var asset = await _context.Assets.FindAsync(exisitingRequest.AssetId);
+                if (asset != null)
+                {
+                    asset.Asset_Status = Models.MultiValues.AssetStatus.OpenToRequest;
+                    await _asset.UpdateAsset(asset);
+                    await _asset.Save();
+
+                    var allocation = await _context.AssetAllocations
+                       .Where(a => a.AssetId == exisitingRequest.AssetId && a.UserId == exisitingRequest.UserId)
+                       .FirstOrDefaultAsync();
+
+                    if (allocation != null)
+                    {
+                        try
+                        {
+                            await _assetAlloc.DeleteAllocation(allocation.AllocationId);
+                            await _assetAlloc.Save();
+                        }
+                        catch (Exception ex)
+                        {
+                            return BadRequest($"Failed to delete allocation with ID {allocation.AllocationId}: {ex.Message}");
+                        }
+                    }
+
+                    var assetRequest = await _context.AssetRequests
+                        .Where(a => a.AssetId == exisitingRequest.AssetId && a.UserId == exisitingRequest.UserId && a.Request_Status == Models.MultiValues.RequestStatus.Allocated)
+                        .FirstOrDefaultAsync();
+
+                    if (assetRequest != null)
+                    {
+                        try
+                        {
+                            _assetRequest.DeleteAssetRequest(assetRequest.AssetReqId);
+                            await _asset.Save();
+                        }
+                        catch (Exception ex)
+                        {
+                            return BadRequest($"Failed to delete AssetRequest with ID {assetRequest.AssetReqId}: {ex.Message}");
+                        }
+                    }
+
+                   
+                }
+            }
             try
             {
-                //await _context.SaveChangesAsync();
-                await _returnRequestRepo.Save();
+               await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -131,11 +163,16 @@ namespace Hexa_Hub.Controllers
         // POST: api/ReturnRequests
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        [Authorize(Roles ="User")]
+        [Authorize(Roles ="Employee")]
         public async Task<ActionResult<ReturnRequest>> PostReturnRequest(ReturnRequest returnRequest)
         {
-            //_context.ReturnRequests.Add(returnRequest);
-            //await _context.SaveChangesAsync();
+            var loggedInUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+            var userhasAsset = await _returnRequestRepo.UserHasAsset(loggedInUserId);
+            if (!userhasAsset)
+            {
+                return BadRequest();
+            }
+            returnRequest.UserId = loggedInUserId;
             _returnRequestRepo.AddReturnRequest(returnRequest);
             await _returnRequestRepo.Save();
 
@@ -144,29 +181,28 @@ namespace Hexa_Hub.Controllers
 
         // DELETE: api/ReturnRequests/5
         [HttpDelete("{id}")]
-        [Authorize(Roles = "User")]
+        [Authorize(Roles = "Employee")]
         public async Task<IActionResult> DeleteReturnRequest(int id)
         {
-            //var returnRequest = await _context.ReturnRequests.FindAsync(id);
-            //if (returnRequest == null)
-            //{
-            //    return NotFound();
-            //}
-
-            //_context.ReturnRequests.Remove(returnRequest);
-            //await _context.SaveChangesAsync();
-
-            //return NoContent();
             try
             {
+                var loggedInUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+                var returnRequest = await _returnRequestRepo.GetReturnRequestById(id);
+                if (returnRequest == null)
+                {
+                    return NotFound();
+                }
+                if (returnRequest.UserId != loggedInUserId)
+                {
+                    return Forbid();
+                }
                 await _returnRequestRepo.DeleteReturnRequest(id);
                 await _returnRequestRepo.Save();
+
                 return NoContent();
             }
             catch (Exception)
             {
-                if (id == null)
-                    return NotFound();
                 return BadRequest();
             }
         }
